main: move target specification into a separate package
