+++ compileopts/target_test.go
@@ -1,4 +1,4 @@
-package main
+package compileopts
 
 import "testing"
