+++ main_test.go
@@ -14,6 +14,7 @@ import (
 	"sort"
 	"testing"
 
+	"github.com/tinygo-org/tinygo/compileopts"
 	"github.com/tinygo-org/tinygo/loader"
 )
 
@@ -142,7 +143,7 @@ func runTest(path, tmpdir string, target string, t *testing.T) {
 	if target == "" {
 		cmd = exec.Command(binary)
 	} else {
-		spec, err := LoadTarget(target)
+		spec, err := compileopts.LoadTarget(target)
 		if err != nil {
 			t.Fatal("failed to load target spec:", err)
 		}
